+++ rocks/__init__.py
@@ -22,9 +22,9 @@ INDEX = {}
 
 # ------
 # Welcome to rocks
-__version__ = "1.4.23"
+__version__ = "1.4.24"
 
-GREETING = fr"""
+GREETING = rf"""
                 _
                | |
  _ __ ___   ___| | _____